+++ news-bundle/CHANGELOG.md
@@ -2,4 +2,5 @@
 
 ### DEV
 
+ * Support sorting enclosures (see contao/calendar-bundle#16).
  * Allow to sort news items by date, headline or randomly (see #13).