+++ news-bundle/CHANGELOG.md
@@ -2,5 +2,6 @@
 
 ### DEV
 
+ * Add the teaser image as media:content in the RSS/Atom feeds (see #7).
  * Support sorting enclosures (see contao/calendar-bundle#16).
  * Allow to sort news items by date, headline or randomly (see #13).